requirements.txt
streamlit
pandas
scikit-learn
joblib
plotly
matplotlib
imapclient
pyzmail36